Add files via upload
This C++ application demonstrates the activities of user-defined constructors and destructors. The constructor is applied when an object is created, and the destructor is applied when the object is destroyed.
Key points:
Constructor: When the object t is created, prints "Constructor executed".
Destructor: Prints "Destructor called" when object t is dest at the end of main().