Add files via upload
Description
This C++ program illustrates the application of default and parameterized constructors as well as methods of a class. The Rectangle class contains methods to calculate area and print dimensions of rectangle
ey Points:
Default Constructor: Initializes the rectangle with length = 0 and breadth = 0.
Parameterized Constructor: Initializes the rectangle with length and breadth given by user.
Method area(): Calculates the area of rectangle (length * breadth).
Method display(): Displays the dimensions of the rectangle.